requirements.txt
streamlit
pytesseract
numpy
Pillow
opencv-python
